Refactor : detach handsUpBtn directory - add Chatting dir - add mainAlert dir - add Note dir
