--- app/src/main/java/com/softsquared/template/kotlin/src/main/handsUpBtn/PostData.kt
@@ -1,8 +1,0 @@
-package com.softsquared.template.kotlin
-
-data class PostData(
-    val id : String = "",
-    val name : String = "" ,
-    val postContent : String = ""
-)
-
