+++ app/src/main/java/com/softsquared/template/kotlin/src/main/mainAlert/FirstFragment.kt
@@ -1,0 +1,36 @@
+package com.softsquared.template.kotlin.src.main.mainAlert
+
+import android.os.Bundle
+import android.util.Log
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.google.firebase.database.ChildEventListener
+import com.google.firebase.database.DataSnapshot
+import com.google.firebase.database.DatabaseError
+import com.google.firebase.database.ktx.database
+import com.google.firebase.ktx.Firebase
+import com.softsquared.template.kotlin.databinding.FragmentFirstBinding
+
+
+/**
+ * A simple [Fragment] subclass as the default destination in the navigation.
+ */
+class FirstFragment : Fragment() {
+
+    private lateinit var viewBinding: FragmentFirstBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+
+        viewBinding = FragmentFirstBinding.inflate(layoutInflater)
+
+        return  viewBinding.root
+    }
+
+}
