+++ app/src/main/java/com/softsquared/template/kotlin/src/main/mainHome/listFragment.kt
@@ -1,7 +1,6 @@
 package com.softsquared.template.kotlin.src.main.mainHome
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -13,8 +12,6 @@ import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
-import com.softsquared.template.kotlin.PostData
-import com.softsquared.template.kotlin.Postdapter
 import com.softsquared.template.kotlin.R
