+++ app/src/main/res/layout/activity_chat.xml
@@ -5,6 +5,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ChatActivity">
+    tools:context=".src.main.Chatting.ChatActivity">
 
 </androidx.appcompat.widget.LinearLayoutCompat>
