+++ app/src/main/res/layout/activity_main_alert.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".src.main.handsUpBtn.MainActivityAlert">
+    tools:context=".src.main.mainAlert.MainActivityAlert">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/navigation"
