+++ app/src/main/res/layout/activity_main_uchat.xml
@@ -8,7 +8,7 @@
     android:paddingTop="20dp"
     android:paddingRight="10dp"
     android:paddingBottom="10dp"
-    tools:context=".ChatActivity">
+    tools:context=".src.main.Chatting.ChatActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_view"
