+++ app/src/main/res/layout/fragment_second.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SecondFragment">
+    tools:context=".src.main.mainAlert.SecondFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_note"
